added commits

+++ src/Utils.py
@@ -6,7 +6,7 @@ def get_airplane_params(df:pd.DataFrame) -> dict:
     for index, row in df.iterrows():
         airplane_params[row["var_name"]] = float(row["var_val"])
 
-    airplane_params["mass"] = float(10) # kg
+    airplane_params["mass"] = float(12) # kg
     
     return airplane_params
